requirements.txt
django==4.2.6
python-dotenv==1.0.0
djangorestframework==3.14.0
requests==2.31.0
aiogram==3.1.1
pydantic==2.1.1
geopy==2.4.0
